+++ x/btcstaking/types/expected_keepers.go
@@ -12,7 +12,7 @@ import (
 type BTCLightClientKeeper interface {
 	GetBaseBTCHeader(ctx context.Context) *btclctypes.BTCHeaderInfo
 	GetTipInfo(ctx context.Context) *btclctypes.BTCHeaderInfo
-	GetHeaderByHash(ctx context.Context, hash *bbn.BTCHeaderHashBytes) *btclctypes.BTCHeaderInfo
+	GetHeaderByHash(ctx context.Context, hash *bbn.BTCHeaderHashBytes) (*btclctypes.BTCHeaderInfo, error)
 }
 
 type BtcCheckpointKeeper interface {